# Refactor and rewrite with speedups and other improvements