--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -6,11 +6,11 @@ pub fn criterion_benchmark(c: &mut Criterion) {
     let open_tb = || TwoBitFile::open_and_read("assets/foo.2bit").unwrap();
     c.bench_function("sequence", |b| {
         let mut tb = open_tb();
-        b.iter(|| black_box(tb.sequence("chr1", 24..74).unwrap().len()))
+        b.iter(|| black_box(tb.read_sequence("chr1", 24..74).unwrap().len()))
     });
     c.bench_function("full_sequence", |b| {
         let mut tb = open_tb();
-        b.iter(|| black_box(tb.sequence("chr1", ..).unwrap().len()))
+        b.iter(|| black_box(tb.read_sequence("chr1", ..).unwrap().len()))
     });
 }