--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -14,6 +14,10 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         let mut tb = open_tb();
         b.iter(|| black_box(tb.read_sequence("chr1", ..).unwrap().len()))
     });
+    c.bench_function("full_soft", |b| {
+        let mut tb = open_tb().enable_softmask(true);
+        b.iter(|| black_box(tb.read_sequence("chr1", ..).unwrap().len()))
+    });
 }
 
 criterion_group!(benches, criterion_benchmark);