--- a/src/block.rs
+++ b/src/block.rs
@@ -16,8 +16,9 @@ pub struct Block {
 
 impl Block {
     /// Create a new block
-    pub fn new(start: Field, length: Field) -> Block {
-        Block { start, length }
+    #[must_use]
+    pub const fn new(start: Field, length: Field) -> Self {
+        Self { start, length }
     }
 
     /// Determine an overlap with another block
@@ -25,7 +26,8 @@ impl Block {
     /// Convenience function to determine overlaps between sequence regions.
     ///
     /// Note: Blocks of the same type should usually not overlap in practise.
-    pub fn overlap(&self, other: &Block) -> Option<Range<usize>> {
+    #[must_use]
+    pub fn overlap(&self, other: &Self) -> Option<Range<usize>> {
         let start = max(self.start, other.start);
         let end = min(self.start + self.length, other.start + other.length);
         if start < end {
@@ -42,19 +44,22 @@ mod tests {
 
     #[test]
     fn test_overlap() {
-    let a = Block::new(5, 10);
-    let b = Block::new(10, 10);
-    let c = Block::new(15, 10);
-    let d = Block::new(100, 25);
-    let e = Block::new(90, 40);
-    assert_eq!(a.overlap(&b), Some(10 .. 15));
-    assert_eq!(a.overlap(&c), None);
-    assert_eq!(b.overlap(&c), Some(15 .. 20));
-    assert_eq!(a.overlap(&a), Some(5 .. 15));
-    assert_eq!(a.overlap(&d), None);
-    assert_eq!(d.overlap(&a), None);
-    assert_eq!(d.overlap(&e), Some(100..125));
-    assert_eq!(e.overlap(&d), Some(100..125));
-    assert_eq!(Block::new(67920552, 300000).overlap(&Block::new(67859142, 61415 )), Some(67920552 .. (67859142 + 61415)));
+        let a = Block::new(5, 10);
+        let b = Block::new(10, 10);
+        let c = Block::new(15, 10);
+        let d = Block::new(100, 25);
+        let e = Block::new(90, 40);
+        assert_eq!(a.overlap(&b), Some(10..15));
+        assert_eq!(a.overlap(&c), None);
+        assert_eq!(b.overlap(&c), Some(15..20));
+        assert_eq!(a.overlap(&a), Some(5..15));
+        assert_eq!(a.overlap(&d), None);
+        assert_eq!(d.overlap(&a), None);
+        assert_eq!(d.overlap(&e), Some(100..125));
+        assert_eq!(e.overlap(&d), Some(100..125));
+        assert_eq!(
+            Block::new(67920552, 300000).overlap(&Block::new(67859142, 61415)),
+            Some(67920552..(67859142 + 61415))
+        );
     }
 }