--- a/src/counts.rs
+++ b/src/counts.rs
@@ -2,8 +2,10 @@
 //!
 //! This module makes no assumptions about softmasked sequences. An "t" is the same as a "T".
 
+use crate::error::{Error, Result};
+
 /// Number of bases of each type in a sequence
-#[derive(PartialEq, Debug, Default)]
+#[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub struct BaseCounts {
     pub a: usize,
     pub c: usize,
@@ -17,10 +19,35 @@ impl BaseCounts {
     pub const fn sum(&self) -> usize {
         self.a + self.c + self.g + self.t + self.n
     }
+
+    #[must_use]
+    pub fn percentages(&self) -> BasePercentages {
+        let mut result = BasePercentages::default();
+        let sum = self.sum() as f64;
+        result.a = (self.a as f64) / sum;
+        result.c = (self.c as f64) / sum;
+        result.g = (self.g as f64) / sum;
+        result.t = (self.t as f64) / sum;
+        result.n = (self.n as f64) / sum;
+        result
+    }
+
+    #[inline]
+    pub(crate) fn update(&mut self, nuc: u8) -> Result<()> {
+        match nuc {
+            b'A' | b'a' => self.a += 1,
+            b'C' | b'c' => self.c += 1,
+            b'G' | b'g' => self.g += 1,
+            b'T' | b't' => self.t += 1,
+            b'N' => self.n += 1,
+            _ => return Err(Error::BadNucleotide(char::from(nuc))),
+        }
+        Ok(())
+    }
 }
 
 /// Percentages of bases of each type in a sequence
-#[derive(PartialEq, Debug, Default)]
+#[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub struct BasePercentages {
     pub a: f64,
     pub c: f64,
@@ -31,13 +58,6 @@ pub struct BasePercentages {
 
 impl From<BaseCounts> for BasePercentages {
     fn from(counts: BaseCounts) -> Self {
-        let mut result = Self::default();
-        let sum = counts.sum() as f64;
-        result.a = (counts.a as f64) / sum;
-        result.c = (counts.c as f64) / sum;
-        result.g = (counts.g as f64) / sum;
-        result.t = (counts.t as f64) / sum;
-        result.n = (counts.n as f64) / sum;
-        result
+        counts.percentages()
     }
 }