--- a/src/error.rs
+++ b/src/error.rs
@@ -1,6 +1,5 @@
 //! Error handling and descriptions
 
-use std::convert::From;
 use std::fmt;
 use std::io;
 use std::string::FromUtf8Error;
@@ -35,12 +34,12 @@ impl std::error::Error for Error {}
 
 impl From<io::Error> for Error {
     fn from(io: io::Error) -> Self {
-        Error::IO(io)
+        Self::IO(io)
     }
 }
 
 impl From<FromUtf8Error> for Error {
     fn from(utf_error: FromUtf8Error) -> Self {
-        Error::FileFormat(format!("{}", utf_error))
+        Self::FileFormat(format!("{}", utf_error))
     }
 }