--- a/src/types.rs
+++ b/src/types.rs
@@ -1,4 +1,3 @@
 //! Central type definitions
 
 pub type Field = u32;
-pub type FileIndex = u32;