Hi! I've spent a bit of time on reworking this crate (as I surprisingly found it to be the slowest part in the processing toolchain which got me to dig a bit deeper to figure out why's that).

Since there's so many changes and it's pretty much a backwards-incompatible rewrite (although the tests are kept and a few more are added), here's a few options, all of which are fine with me and understandable:

- This gets merged and you keep on being a sole maintainer (or add me as comaintainer if you want)
- You aren't able or don't wish to maintain this crate any longer and want to pass it on along with its name
- Something inbetween the previous two options where the maintenance and rights are shared in some way, whichever way you'd prefer
- You don't want to merge this and want to retain your crate, in which case I would have to publish another one since I need it working and published asap

Here's a list of some of the changes:

- Parsing algorithm is rewritten from scratch and is much faster
- Sequence mask replacing algorithm is rewritten from scratch and is much faster as well
- Docs are improved, many docstrings are rewritten
- Reader is no longer based on `BufReader<File>` only, but can be placed on top of any reader
- Built-in support for in-memory-cached files which provides lightning-fast decoding
- Remove half of the API by using Rust ranges everywhere to slice sequences when decoding
- Many other API changes, making it backwards-incompatible but nicer (in my opinion) and more Rust-like (but less py2bit-like, whether that's a good or bad thing)
- The entire codebase is now conforming to the latest clippy
- Added GitHub actions workflows so that the tests/clippy run automatically
- Added simple benchmarks that can be run directly via cargo

Timings comparison:

```
master:
    sequence                time:   [20.588 us 20.774 us 20.976 us]
    full_sequence           time:   [21.221 us 21.339 us 21.464 us]
PR (from disk):
    sequence                time:   [1.5980 us 1.6061 us 1.6150 us] (~13x)
    full_sequence           time:   [1.6980 us 1.7057 us 1.7137 us]
PR (from memory):
    sequence                time:   [105.97 ns 106.27 ns 106.62 ns] (~196x)
    full_sequence           time:   [171.51 ns 178.67 ns 185.83 ns]
```