@jbethune 2021-11-18
Wow. Thanks for the pull request! Looks promising at first sight. But I will need a bit of time to review everything and make the decisions you mentioned.