@aldanor 2021-11-18
@jbethune I've rewritten `apply_masks()` with no unsafe and in this case it's not slower (but actually a few percent faster, at least on my machine) - so, good call!