@aldanor 2021-11-18
@jbethune I've rewritten the decoder once again 🤦 

- On a local machine, went from 119ns to 98ns in the partial bench (approximately 20% speedup)
- No more unsafe in the main loop
- Much less code
- Uses a preparsed "database" of quadruples (see previous commit)