@jbethune 2021-11-19
> Just a note: I've received an email re: crate co-ownership from https://crates.io, but it seems like the invitation has expired somehow, or something went wrong, maybe try again?

Is it still not working? [You are listed in the owner's section](https://crates.io/crates/twobit)