@aldanor 2021-11-19
> > Just a note: I've received an email re: crate co-ownership from https://crates.io, but it seems like the invitation has expired somehow, or something went wrong, maybe try again?
> 
> Is it still not working? [You are listed in the owner's section](https://crates.io/crates/twobit)

I think it works now, maybe something was lagging yesterday and it didn't update yet back then.

So what's the plan then? 

- I'll add a brief changelog, it's a good thing to have.
- If everything looks fine, I can try merging this in (and we need to ensure that GitHub Actions will then kick off in this repo and start running tests, first of all)
- Add 0.2.0 git tag and cargo publish that commit, then add a github release (by the way - I can't seem to edit description of the github repo etc - perhaps there's some more checkboxes you might have to set in the repository settings if you want to)
- One more thing - would you like the license to be MIT or MIT + Apache2?